+++ src/infra/repositories/PetRepoFile.cpp
@@ -138,7 +138,7 @@ Dog PetRepository::getDogInfo(const string &petCode) {
 
     description = trimQuotes(description); // Xóa dấu ngoặc kép
 
-    dog = Dog(code, name, breed, stoi(age), stof(price), stoi(energy),
+    dog = Dog(code, name, breed, stoi(age), stol(price), stoi(energy),
               description);
     return dog;
 }
@@ -166,7 +166,7 @@ Cat PetRepository::getCatInfo(const string &petCode) {
 
     description = trimQuotes(description); // Xóa dấu ngoặc kép
 
-    cat = Cat(code, name, breed, stoi(age), stof(price), fur, description);
+    cat = Cat(code, name, breed, stoi(age), stol(price), fur, description);
     return cat;
 }
 
@@ -196,7 +196,7 @@ LinkedList<Cat> PetRepository::getAllCatInfo() {
 
         description = trimQuotes(description); // Xóa dấu ngoặc kép
 
-        Cat cat(code, name, breed, stoi(age), stof(price), fur, description);
+        Cat cat(code, name, breed, stoi(age), stol(price), fur, description);
         cats.pushBack(cat);
     }
     file.close();
@@ -229,7 +229,7 @@ LinkedList<Dog> PetRepository::getAllDogInfo() {
 
         description = trimQuotes(description); // Xóa dấu ngoặc kép
 
-        Dog dog(code, name, breed, stoi(age), stof(price), stoi(energy),
+        Dog dog(code, name, breed, stoi(age), stol(price), stoi(energy),
                 description);
         dogs.pushBack(dog);
     }
@@ -264,7 +264,7 @@ LinkedList<Cat> PetRepository::getAllCatInfoAvailable() {
         description = trimQuotes(description); // Xóa dấu ngoặc kép
 
         if (status == "1") {
-            Cat cat(code, name, breed, stoi(age), stof(price), fur, description);
+            Cat cat(code, name, breed, stoi(age), stol(price), fur, description);
             cats.pushBack(cat);
         }
     }
@@ -299,7 +299,7 @@ LinkedList<Dog> PetRepository::getAllDogInfoAvailable() {
         description = trimQuotes(description); // Xóa dấu ngoặc kép
 
         if (status == "1") {
-            Dog dog(code, name, breed, stoi(age), stof(price), stoi(energy),
+            Dog dog(code, name, breed, stoi(age), stol(price), stoi(energy),
                     description);
             dogs.pushBack(dog);
         }
